move client-side scripts from '/scripts' to '/client/utils'
