+++ client/utils/threshold.js
@@ -1,0 +1,6 @@
+const threshold = value => {
+  // return white or black
+  return value > 0x15 ? 0xff : 0;
+};
+
+export default threshold;
