--- scripts/index.js
@@ -1,10 +1,0 @@
-export { default as connectToEasyRTC } from './connectToEasyRTC';
-export { default as drawVideo } from './drawVideo';
-export { default as blend } from './blend';
-export { default as differenceAccuracy } from './differenceAccuracy';
-export { default as fastAbs } from './fastAbs';
-export { default as threshold } from './threshold';
-export { default as checkAreas } from './checkAreas';
-export { default as socketEvents } from './socketEvents';
-export { default as pirateDictionary } from './pirateDictionary';
-export { default as shipNameGenerator } from './shipNameGenerator';
